refactor: extract pane state module
